requirements.txt
django==3.1.1
python-dotenv==0.12.0
pytest-black==0.3.7
psycopg2==2.9.1